Extract DuckDB-side type stubs to dsgrid/ibis/_duckdb_shims.py
Pulls the per-type stub classes (BooleanType, IntegerType, etc., plus
Row, SparkConf, AnalysisException, StructField, StructType, and the
_UnsupportedSparkFunctions / F proxy) out of session.py and into a new
leaf module. session.py's else-branch shrinks from ~90 lines of inline
class definitions to a single import block.

_SparkSessionBuilder and SparkSession remain inline in session.py
because they reach back into get_runtime_session() and would otherwise
require a circular import. _DuckDBConf, _DuckDBCatalog, _DuckDBReader
also stay for the same reason and because _DuckDBReader uses the
schema helpers defined further down the file.

No behavior change: the stubs are bit-for-bit the same, just relocated.
